Replaced boxart for "Adventures to Go! (USA)"
I'm making a separate PR for this because the new image is of smaller resolution then the current. Interestingly the current boxart has Pegi 3 on it, but I could find no evidence that it was ever released with a rating other than Pegi 12. (This was one of those that have a EU cover for the USA release, and the EU release has no cover.)