Delete Final Fantasy III (USA).png
Replaced with ever so slightly higher quality version of Andrea's Japanese box art. Hawaka's version needs to be deleted. It is auto downloading to playlists, and having two images under the same name is not proper. His version is just a random piece of art without an actual box, and that is what auto DL'd onto my playlist. Very confusing. It needs to be renamed or used as alternative. Both of these are irrelevant now.